chore: upgrade provider from `2.8.0` to version `2.8.1`
