+++ package.json
@@ -149,7 +149,7 @@
   "cdktf": {
     "provider": {
       "name": "registry.terraform.io/upcloudltd/upcloud",
-      "version": "2.8.0"
+      "version": "2.8.1"
     }
   },
   "resolutions": {
