+++ src/data-upcloud-hosts/index.ts
@@ -116,7 +116,7 @@ export class DataUpcloudHosts extends cdktf.TerraformDataSource {
       terraformResourceType: 'upcloud_hosts',
       terraformGeneratorMetadata: {
         providerName: 'upcloud',
-        providerVersion: '2.8.0',
+        providerVersion: '2.8.1',
         providerVersionConstraint: '~> 2.4'
       },
       provider: config.provider,
