+++ src/data-upcloud-ip-addresses/index.ts
@@ -146,7 +146,7 @@ export class DataUpcloudIpAddresses extends cdktf.TerraformDataSource {
       terraformResourceType: 'upcloud_ip_addresses',
       terraformGeneratorMetadata: {
         providerName: 'upcloud',
-        providerVersion: '2.8.0',
+        providerVersion: '2.8.1',
         providerVersionConstraint: '~> 2.4'
       },
       provider: config.provider,
