+++ src/data-upcloud-kubernetes-cluster/index.ts
@@ -44,7 +44,7 @@ export class DataUpcloudKubernetesCluster extends cdktf.TerraformDataSource {
       terraformResourceType: 'upcloud_kubernetes_cluster',
       terraformGeneratorMetadata: {
         providerName: 'upcloud',
-        providerVersion: '2.8.0',
+        providerVersion: '2.8.1',
         providerVersionConstraint: '~> 2.4'
       },
       provider: config.provider,
