+++ src/data-upcloud-zone/index.ts
@@ -48,7 +48,7 @@ export class DataUpcloudZone extends cdktf.TerraformDataSource {
       terraformResourceType: 'upcloud_zone',
       terraformGeneratorMetadata: {
         providerName: 'upcloud',
-        providerVersion: '2.8.0',
+        providerVersion: '2.8.1',
         providerVersionConstraint: '~> 2.4'
       },
       provider: config.provider,
