+++ src/firewall-rules/index.ts
@@ -524,7 +524,7 @@ export class FirewallRules extends cdktf.TerraformResource {
       terraformResourceType: 'upcloud_firewall_rules',
       terraformGeneratorMetadata: {
         providerName: 'upcloud',
-        providerVersion: '2.8.0',
+        providerVersion: '2.8.1',
         providerVersionConstraint: '~> 2.4'
       },
       provider: config.provider,
