+++ src/index.ts
@@ -2,6 +2,7 @@
 export * as firewallRules from './firewall-rules';
 export * as floatingIpAddress from './floating-ip-address';
 export * as kubernetesCluster from './kubernetes-cluster';
+export * as kubernetesNodeGroup from './kubernetes-node-group';
 export * as loadbalancer from './loadbalancer';
 export * as loadbalancerBackend from './loadbalancer-backend';
 export * as loadbalancerDynamicBackendMember from './loadbalancer-dynamic-backend-member';
