+++ src/kubernetes-node-group/README.md
@@ -1,0 +1,3 @@
+# `upcloud_kubernetes_node_group`
+
+Refer to the Terraform Registory for docs: [`upcloud_kubernetes_node_group`](https://www.terraform.io/docs/providers/upcloud/r/kubernetes_node_group).
