+++ src/loadbalancer-backend/index.ts
@@ -394,7 +394,7 @@ export class LoadbalancerBackend extends cdktf.TerraformResource {
       terraformResourceType: 'upcloud_loadbalancer_backend',
       terraformGeneratorMetadata: {
         providerName: 'upcloud',
-        providerVersion: '2.8.0',
+        providerVersion: '2.8.1',
         providerVersionConstraint: '~> 2.4'
       },
       provider: config.provider,
