+++ src/loadbalancer-frontend-tls-config/index.ts
@@ -60,7 +60,7 @@ export class LoadbalancerFrontendTlsConfig extends cdktf.TerraformResource {
       terraformResourceType: 'upcloud_loadbalancer_frontend_tls_config',
       terraformGeneratorMetadata: {
         providerName: 'upcloud',
-        providerVersion: '2.8.0',
+        providerVersion: '2.8.1',
         providerVersionConstraint: '~> 2.4'
       },
       provider: config.provider,
