+++ src/loadbalancer-frontend/index.ts
@@ -276,7 +276,7 @@ export class LoadbalancerFrontend extends cdktf.TerraformResource {
       terraformResourceType: 'upcloud_loadbalancer_frontend',
       terraformGeneratorMetadata: {
         providerName: 'upcloud',
-        providerVersion: '2.8.0',
+        providerVersion: '2.8.1',
         providerVersionConstraint: '~> 2.4'
       },
       provider: config.provider,
