+++ src/loadbalancer-manual-certificate-bundle/index.ts
@@ -66,7 +66,7 @@ export class LoadbalancerManualCertificateBundle extends cdktf.TerraformResource
       terraformResourceType: 'upcloud_loadbalancer_manual_certificate_bundle',
       terraformGeneratorMetadata: {
         providerName: 'upcloud',
-        providerVersion: '2.8.0',
+        providerVersion: '2.8.1',
         providerVersionConstraint: '~> 2.4'
       },
       provider: config.provider,
