+++ src/loadbalancer-resolver/index.ts
@@ -91,7 +91,7 @@ export class LoadbalancerResolver extends cdktf.TerraformResource {
       terraformResourceType: 'upcloud_loadbalancer_resolver',
       terraformGeneratorMetadata: {
         providerName: 'upcloud',
-        providerVersion: '2.8.0',
+        providerVersion: '2.8.1',
         providerVersionConstraint: '~> 2.4'
       },
       provider: config.provider,
