+++ src/managed-database-logical-database/index.ts
@@ -66,7 +66,7 @@ export class ManagedDatabaseLogicalDatabase extends cdktf.TerraformResource {
       terraformResourceType: 'upcloud_managed_database_logical_database',
       terraformGeneratorMetadata: {
         providerName: 'upcloud',
-        providerVersion: '2.8.0',
+        providerVersion: '2.8.1',
         providerVersionConstraint: '~> 2.4'
       },
       provider: config.provider,
