+++ src/object-storage/index.ts
@@ -190,7 +190,7 @@ export class ObjectStorage extends cdktf.TerraformResource {
       terraformResourceType: 'upcloud_object_storage',
       terraformGeneratorMetadata: {
         providerName: 'upcloud',
-        providerVersion: '2.8.0',
+        providerVersion: '2.8.1',
         providerVersionConstraint: '~> 2.4'
       },
       provider: config.provider,
