+++ src/router/index.ts
@@ -48,7 +48,7 @@ export class Router extends cdktf.TerraformResource {
       terraformResourceType: 'upcloud_router',
       terraformGeneratorMetadata: {
         providerName: 'upcloud',
-        providerVersion: '2.8.0',
+        providerVersion: '2.8.1',
         providerVersionConstraint: '~> 2.4'
       },
       provider: config.provider,
