+++ src/server-group/index.ts
@@ -73,7 +73,7 @@ export class ServerGroup extends cdktf.TerraformResource {
       terraformResourceType: 'upcloud_server_group',
       terraformGeneratorMetadata: {
         providerName: 'upcloud',
-        providerVersion: '2.8.0',
+        providerVersion: '2.8.1',
         providerVersionConstraint: '~> 2.4'
       },
       provider: config.provider,
