+++ src/tag/index.ts
@@ -60,7 +60,7 @@ export class Tag extends cdktf.TerraformResource {
       terraformResourceType: 'upcloud_tag',
       terraformGeneratorMetadata: {
         providerName: 'upcloud',
-        providerVersion: '2.8.0',
+        providerVersion: '2.8.1',
         providerVersionConstraint: '~> 2.4'
       },
       provider: config.provider,
