+++ src/version.json
@@ -1,3 +1,3 @@
 {
-  "registry.terraform.io/upcloudltd/upcloud": "2.8.0"
+  "registry.terraform.io/upcloudltd/upcloud": "2.8.1"
 }
